pyproject.toml
[tool.poetry]
name = "polls"
version = "0.1.0"
description = "Polls with DRF and Django"
authors = ["mbrav <mbrav@protonmail.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.10"
Django = "<4"
pytest = "^6.2"
djangorestframework = "^3.13.1"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[tool.pytest.ini_options]
minversion = "6.2"
addopts = "--docker-compose=docker/docker-compose.yml"
testpaths = [
    "tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

